pyproject.toml
[project]
name = "cat-tv"
version = "2.0.0"
description = "Raspberry Pi Cat TV - YouTube stream player with web interface for cat entertainment"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "yt-dlp>=2023.12.30",
    "python-vlc>=3.0.20123",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.23",
    "flask-socketio>=5.3.0",
    "python-socketio>=5.10.0",
]

[project.scripts]
cat-tv = "cat_tv.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ["py39"]